Simple Dice
Experimenting with "coding" via ChatGPT

I want a single, simple page website.  On the left side there is a button which reads "Roll," and on the right side of the screen is a text box.  Every time a user clicks the "roll" button, a six sided dice is rolled, and the value gets put into the text box.  This should be done with HTML and Javascript, there should be no server side code

Looks good!  Add an input box labeled Number of Dice.  Now I would like an updated rollDice function.  It takes an integer, and rolls that number of dice.  The function gets its value from the Number of Dice input box.  We are also changing the type of dice that gets rolled.  Instead of being a six-sided dice with the numbers 1, 2, 3, 4, 5, 6, we have a dice with the numbers 0, 0, 1, 1, 2, 2.  When we finish rolling the dice, the sum of the dice values should be output.